INKLB-001 Sessions manager uses weak_ptr to session
It should not own session
When session is called, it should be shared_ptr => it wouldn't be deleted
